% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot geographical coordinates of fatal injury locations}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a valid state identifier, coercible to an integer}

\item{year}{a scalar, coercible to an integer}
}
\value{
\code{NULL}
}
\description{
\code{fars_map_state} plots the geograhpical coordinates
of fatal injuries within a state given a year
}
\details{
The plotted area is restricted to states with a longitude
> 900 and latitude > 90
}
\examples{
\dontrun{
fars_map_state(01, 2013)
}
}
\seealso{
\code{\link[dplyr]{filter}}
 \code{\link[maps]{map}}
 \code{\link[graphics]{points}}
}
