% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to a csv file}
}
\value{
Returns the data in a tibble format
}
\description{
This functions is a helper for reading FARS data in csv format.
It returns the data in a \code{\link[dplyr]{tbl_df}}
}
\examples{
\dontrun{
fars_read("filename.csv")
}
}
