% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data of multiple years into a list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector that can be coerced to an integer vector
of valid years}
}
\value{
Returns a list with each element being the data for that
year in a \code{tbl_df}
}
\description{
This function is a helper for reading FARS data in csv format.
It returns the data in a list with each element being a tibble for 
the specified years
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
}
