% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fatal injury counts by (year, month)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector that can be coerced into an integer vector
of valid years}
}
\value{
a \code{data.frame} with monthly counts of fatal injuries
for each year
}
\description{
Reads in FARS csv files for multiple years and
summarises the monthly count of fatal injuries suffered in
traffic crashes
}
\examples{
\dontrun{
fars_summarize_year(2013:2015)
}
}
\seealso{
\code{\link[dplyr]{bind}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
 \code{\link[tidyr]{spread}}
}
