% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Helper for creating FARS data filenames}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a scalar or vector which can be coerced integer values}
}
\value{
a character vector with filenames according to the year
}
\description{
\code{make_filename} is a helper for creating the filenames using
year values for the FARS data
}
\examples{
\dontrun{
make_filename("2013")
make_filename(2013)
make_filename(2013:2015)
}
}
